% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ParameterCheck.R
\docType{package}
\name{ParameterCheck}
\alias{ParameterCheck}
\alias{ParameterCheck-package}
\title{Improved Parameter Check Communication}
\description{
The typical process of checking arguments in functions is iterative.  In
this process, an error may be returned and the user may fix it only to receive
another error on a different argument.  'ParameterCheck 'facilitates a more helpful
way to perform argument checks allowing the programmer to run all of the checks
and then return all of the errors and warnings in a single message.
}
\examples{
\dontrun{
#* This example is taken from the discussion of argument checking at
#* http://www.r-bloggers.com/programming-with-r---checking-function-arguments/
cylinder.volume <- function(height, radius){
  PCheck <- newParamCheck()
  PCheck <- addError(missing(height),
                     "A value for 'height' was not provided",
                     PCheck)
  PCheck <- addError(ifelse(!missing(height), height < 0, FALSE),
                     "'height' must be a non-negative number",
                     PCheck)
  PCheck <- addError(missing(radius),
                     "A value for 'radius' was not provided",
                     PCheck)
  PCheck <- addError(ifelse(!missing(radius), radius < 0, FALSE),
                     "'radius' must be a non-negative number",
                     PCheck)

  PCheck <- addWarning(ifelse(!missing(height) & !missing(radius),
                              height < radius, FALSE),
                       "When 'height' < 'radius', you have a short, awkward looking cylinder",
                       PCheck)

finishParamCheck(PCheck)

volume <- pi * radius^2 * height
volume
}

#' cylinder.volume()
cylinder.volume(height = -3)
cylinder.volume(height = 3, radius = -2)
cylinder.volume(height = 3, radius=2)
cylinder.volume(height = -8, radius = 4)
}
}

